requirements.txt

fastapi>=0.88,<0.100
uvicorn[standard]>=0.22
motor>=3.3
pymongo>=4.6
pydantic>=1.10,<2.0
PyJWT>=2.8
bcrypt>=4.1
python-dotenv>=1.0
